+++ src/Elasticsearch/Paginator.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace ApiPlatform\Elasticsearch;
 
 use ApiPlatform\Elasticsearch\Serializer\DocumentNormalizer;
-use ApiPlatform\Metadata\State\Pagination\PaginatorInterface;
+use ApiPlatform\State\Pagination\PaginatorInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
