+++ src/Elasticsearch/State/CollectionProvider.php
@@ -19,9 +19,9 @@ use ApiPlatform\Elasticsearch\Metadata\Document\Factory\DocumentMetadataFactoryI
 use ApiPlatform\Elasticsearch\Paginator;
 use ApiPlatform\Elasticsearch\Util\ElasticsearchVersion;
 use ApiPlatform\Metadata\Operation;
-use ApiPlatform\Metadata\State\Pagination\Pagination;
-use ApiPlatform\Metadata\State\ProviderInterface;
 use ApiPlatform\Metadata\Util\Inflector;
+use ApiPlatform\State\Pagination\Pagination;
+use ApiPlatform\State\ProviderInterface;
 use Elasticsearch\Client;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
