+++ src/Elasticsearch/Tests/Filter/TermFilterTest.php
@@ -176,41 +176,49 @@ class TermFilterTest extends TestCase
                     'property' => 'id',
                     'type' => 'int',
                     'required' => false,
+                    'is_collection' => false,
                 ],
                 'id[]' => [
                     'property' => 'id',
                     'type' => 'int',
                     'required' => false,
+                    'is_collection' => true,
                 ],
                 'name' => [
                     'property' => 'name',
                     'type' => 'string',
                     'required' => false,
+                    'is_collection' => false,
                 ],
                 'name[]' => [
                     'property' => 'name',
                     'type' => 'string',
                     'required' => false,
+                    'is_collection' => true,
                 ],
                 'date' => [
                     'property' => 'date',
                     'type' => \DateTimeInterface::class,
                     'required' => false,
+                    'is_collection' => false,
                 ],
                 'date[]' => [
                     'property' => 'date',
                     'type' => \DateTimeInterface::class,
                     'required' => false,
+                    'is_collection' => true,
                 ],
                 'weird' => [
                     'property' => 'weird',
                     'type' => 'string',
                     'required' => false,
+                    'is_collection' => false,
                 ],
                 'weird[]' => [
                     'property' => 'weird',
                     'type' => 'string',
                     'required' => false,
+                    'is_collection' => true,
                 ],
             ],
             $termFilter->getDescription(Foo::class)
