+++ src/Elasticsearch/Tests/State/CollectionProviderTest.php
@@ -21,7 +21,7 @@ use ApiPlatform\Elasticsearch\Tests\Fixtures\Foo;
 use ApiPlatform\Metadata\Get;
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
 use ApiPlatform\Metadata\Resource\ResourceMetadataCollection;
-use ApiPlatform\Metadata\State\Pagination\Pagination;
+use ApiPlatform\State\Pagination\Pagination;
 use Elasticsearch\Client;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
