+++ src/Elasticsearch/composer.json
@@ -22,6 +22,7 @@
   "require": {
     "php": ">=8.1",
     "api-platform/metadata": "*@dev || ^3.1",
+    "api-platform/state": "*@dev || ^3.1",
     "elasticsearch/elasticsearch": "^7.11.0",
     "symfony/cache": "^6.1",
     "symfony/console": "^6.2",
@@ -67,6 +68,10 @@
     {
       "type": "path",
       "url": "../Metadata"
-    }
+    },
+   {
+     "type": "path",
+     "url": "../State"
+   }
   ]
 }
