--- src/Metadata/State/OptionsInterface.php
@@ -1,20 +1,0 @@
-<?php
-
-/*
- * This file is part of the API Platform project.
- *
- * (c) Kévin Dunglas <dunglas@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace ApiPlatform\Metadata\State;
-
-interface OptionsInterface
-{
-}
-
-class_alias(OptionsInterface::class, \ApiPlatform\State\OptionsInterface::class);
