+++ src/State/OptionsInterface.php
@@ -13,11 +13,6 @@ declare(strict_types=1);
 
 namespace ApiPlatform\State;
 
-interface_exists(\ApiPlatform\Metadata\State\OptionsInterface::class);
-
-/* @phpstan-ignore-next-line */
-if (false) {
-    interface OptionsInterface extends \ApiPlatform\Metadata\State\OptionsInterface
-    {
-    }
+interface OptionsInterface
+{
 }
