--- tests/Elasticsearch/Exception/IndexNotFoundExceptionTest.php
@@ -1,29 +1,0 @@
-<?php
-
-/*
- * This file is part of the API Platform project.
- *
- * (c) Kévin Dunglas <dunglas@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace ApiPlatform\Tests\Elasticsearch\Exception;
-
-use ApiPlatform\Elasticsearch\Exception\IndexNotFoundException;
-use ApiPlatform\Metadata\Exception\ExceptionInterface;
-use PHPUnit\Framework\TestCase;
-
-class IndexNotFoundExceptionTest extends TestCase
-{
-    public function testConstruct(): void
-    {
-        $indexNotFoundException = new IndexNotFoundException();
-
-        self::assertInstanceOf(ExceptionInterface::class, $indexNotFoundException);
-        self::assertInstanceOf(\Throwable::class, $indexNotFoundException);
-    }
-}
